+++ plugin/src/hotReload.lua
@@ -1,5 +1,0 @@
-local function hotReload(module)
-	return require(module:Clone())
-end
-
-return hotReload