+++ plugin/src/testez/TestBootstrap.lua
@@ -49,15 +49,12 @@ local function toStringPath(tablePath)
 	return stringPath
 end
 
-local hotReload = require(script.Parent.Parent.hotReload)
-
 function TestBootstrap:getModulesImpl(root, modules, current)
 	modules = modules or {}
 	current = current or root
 
 	if isSpecScript(current) then
-		-- local method = require(current)
-		local method = hotReload(current) -- modified for TestEZ inspector
+		local method = require(current)
 		local path = getPath(current, root)
 		local pathString = toStringPath(path)
 
@@ -73,11 +70,13 @@ end
 	Find all the ModuleScripts in this tree that are tests.
 ]]
 function TestBootstrap:getModules(root)
+	local clonedRoot = root:Clone() -- modified for TestEZ Companion
+
 	local modules = {}
 
-	self:getModulesImpl(root, modules)
+	self:getModulesImpl(clonedRoot, modules)
 
-	for _, child in ipairs(root:GetDescendants()) do
+	for _, child in ipairs(clonedRoot:GetDescendants()) do
 		self:getModulesImpl(root, modules, child)
 	end