+++ src/PlaceList.d.ts
@@ -1,0 +1,5 @@
+export type PlaceList = {
+	placeId: string;
+	placeName: string;
+	displayName: string;
+}[];